refactor(resume): Split DefaultTheme into modular components
